Set array type to int32

+++ aptools/data_handling/saving.py
@@ -115,7 +115,7 @@ def save_for_csrtrack_fmt1(beam_data, folder_path, file_name, reposition=False,
     if (n_part is not None and n_part < len(q_orig)):
         q_tot = np.sum(q_orig)
         q_part = q_tot/n_part
-        i = np.arange(len(q_orig))
+        i = np.arange(len(q_orig), dtype=np.int32)
         i = np.random.choice(i, size=n_part)
         x_orig = x_orig[i]
         y_orig = y_orig[i]
